+++ src/XrmDefinitelyTyped/IntermediateRepresentation.fs
@@ -64,10 +64,15 @@ module IntermediateRepresentation =
     | Mobile = 5
     | Quick = 6
     | QuickCreate = 7
+    | Dialog = 8
+    | TaskFlowForm = 9
+    | InteractionCentricDashboard = 10
+    | Card = 11
+    | MainInteractionCentric = 12
     | Other = 100
     | MainBackup = 101
     | AppointmentBookBackup = 102
-  
+
   type XrmFormAttribute = string * AttributeType
   type XrmFormControl = string * AttributeType option * ControlType
   type XrmFormTab = string * string * string list